Restore the use of AUTOINCREMENT on SQLite (Fixes #965)
When :auto_increment in the schema hashes was introduced, it
caused some problems in the specs on SQLite because it started
using AUTOINCREMENT in places where it wasn't allowed.  Removing
the use of AUTOINCREMENT fixed things, and I thought it was a
noise word on SQLite, since SQLite integer primary keys are
naturally auto incrementing.  However, AUTOINCREMENT on SQLite
makes it so SQLite will not reuse a previous primary key value,
so removing it changed the behavior.

Restore the old behavior, and change things so that :auto_increment
is removed from the schema hashes used by the schema modification
methods.