Make Database#transaction work in after_commit/after_rollback blocks
This was broken before, because the transaction metadata wasn't
cleared before calling the blocks, so Sequel still thought it was
in the previous transaction, even though that transaction had
either been committed or rolled back.

This breaks backward compatibility for the private API, by removing
the after_transaction_commit and after_transaction_rollback
private methods.